Rename command and resolve TypeScript issues
